pyproject.toml
[project]
name = "post-quantum-mcp"
version = "0.5.0"
description = "MCP server for post-quantum cryptography with hybrid X25519+ML-KEM-768 key exchange and ML-DSA-65 authenticated envelopes. Research/prototyping tooling using liboqs."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Scott Hughes", email = "scottdhughes@users.noreply.github.com"}
]
keywords = [
    "post-quantum",
    "cryptography",
    "mcp",
    "model-context-protocol",
    "liboqs",
    "ml-kem",
    "ml-dsa",
    "kyber",
    "dilithium",
    "sphincs",
    "falcon",
    "quantum-safe",
    "pqc",
    "nist",
    "fips-203",
    "fips-204",
    "hybrid-key-exchange",
    "x25519",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "mcp>=1.6.0,<2.0.0",
    "liboqs-python>=0.10.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/scottdhughes/post-quantum-mcp"
Repository = "https://github.com/scottdhughes/post-quantum-mcp"
Issues = "https://github.com/scottdhughes/post-quantum-mcp/issues"
Documentation = "https://github.com/scottdhughes/post-quantum-mcp#readme"

[project.scripts]
pqc-mcp-server = "pqc_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pqc_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
